restructure files to have a nicer separation
