+++ src/expandableCalendar/asCalendarConsumer.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
 import hoistNonReactStatic from 'hoist-non-react-statics';
-import CalendarContext from './calendarContext';
+import CalendarContext from './Context';
 
 function asCalendarConsumer(WrappedComponent) {
   class CalendarConsumer extends Component {
